Integrate admin-ui java logic into http module
So far we've had 2 RestFilter implementations. 1 in `core` and 1 in
`admin-ui`.
In ES52 the `RestFilter` functionality has been removed. There is a
replacement, but only a single plugin can use that functionality.

This commit moves the "serve JSON on /" functionality from `core` to
`http`. In addition, it moves all logic from `admin-ui` into `http`.
This way we get rid of a bit duplicated guice-injection magic and it
will be easier to merge the two `RestFilter` implementations into one.
