+++ .gitignore
@@ -42,6 +42,7 @@ out/
 app/enterprise/
 app/oss_modules/
 app/plugins/
+app/adminui/
 app/src/main/dist/bin/crate.bat
 app/src/main/dist/bin/crate.in.sh
 app/src/main/resources/crate-build.properties
