--- admin-ui/src/main/java/io/crate/plugin/AdminUIPlugin.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to CRATE Technology GmbH ("Crate") under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  Crate licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- * However, if you have executed another commercial license
- * agreement with Crate these terms will supersede the license
- * and you may use the software solely pursuant to the terms of
- * the relevant commercial agreement.
- */
-
-package io.crate.plugin;
-
-
-import io.crate.module.AdminUIModule;
-import io.crate.rest.action.admin.AdminUIFrontpageAction;
-import org.elasticsearch.common.inject.Module;
-import org.elasticsearch.plugins.ActionPlugin;
-import org.elasticsearch.plugins.Plugin;
-import org.elasticsearch.rest.RestHandler;
-
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * Crate Admin-UI Plugin
- */
-public class AdminUIPlugin extends Plugin implements ActionPlugin {
-
-    @Override
-    public List<Class<? extends RestHandler>> getRestHandlers() {
-        return Collections.singletonList(AdminUIFrontpageAction.class);
-    }
-
-    @Override
-    public Collection<Module> createGuiceModules() {
-        return Collections.singletonList(new AdminUIModule());
-    }
-}
