+++ core/src/main/java/io/crate/plugin/CrateCorePlugin.java
@@ -25,25 +25,19 @@ package io.crate.plugin;
 import io.crate.ClusterIdService;
 import io.crate.metadata.CustomMetaDataUpgraderLoader;
 import io.crate.module.CrateCoreModule;
-import io.crate.rest.CrateRestFilter;
-import io.crate.rest.CrateRestMainAction;
 import org.elasticsearch.cluster.metadata.MetaData;
 import org.elasticsearch.common.component.LifecycleComponent;
 import org.elasticsearch.common.inject.Module;
-import org.elasticsearch.common.settings.Setting;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.IndexModule;
-import org.elasticsearch.plugins.ActionPlugin;
 import org.elasticsearch.plugins.Plugin;
-import org.elasticsearch.rest.RestHandler;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.function.UnaryOperator;
 
-public class CrateCorePlugin extends Plugin implements ActionPlugin {
+public class CrateCorePlugin extends Plugin {
 
     private final Settings settings;
     private final IndexEventListenerProxy indexEventListenerProxy;
@@ -59,19 +53,10 @@ public class CrateCorePlugin extends Plugin implements ActionPlugin {
     }
 
     @Override
-    public List<Setting<?>> getSettings() {
-        return Collections.singletonList(CrateRestFilter.ES_API_ENABLED_SETTING);
-    }
-
-    @Override
     public Collection<Module> createGuiceModules() {
         return Collections.singletonList(new CrateCoreModule(settings, indexEventListenerProxy));
     }
 
-    @Override
-    public List<Class<? extends RestHandler>> getRestHandlers() {
-        return Collections.singletonList(CrateRestMainAction.class);
-    }
 
     @Override
     public UnaryOperator<Map<String, MetaData.Custom>> getCustomMetaDataUpgrader() {
